--- src/components/TimerProgressBar/index.css
@@ -1,32 +1,0 @@
-.container {
-    position: relative; 
-    align-items: center;
-    border-radius: 50%;
-    box-sizing: border-box; 
-}
-.container div {
-    position: absolute;
-    border-radius: 50%;
-    box-sizing: border-box;
-    transition: transform .1s;  
-}
-.container .background {
-    border: 20px solid gray;
-}
-.container .rotate {
-    border: 20px solid #ff5155;
-}
-.container .left {
-    border: 20px solid gray;
-  
-}
-.container .right {
-    transform: rotate(180deg);
-    border: 20px solid #ff5155;
-}
-.textContainer {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    align-items: center;
-}
